Added Access control. Finished Authentication

+++ resources/views/posts/show.blade.php
@@ -15,20 +15,23 @@
         </p>    
         <hr>
         <small>
-            Written on: {{$post->created_at}}
+            Written on: {{$post->created_at}} by <strong>{{$post->user->name}}</strong>
         </small>
         <hr>
-        <div class="row">
-            <div class="col s6">
-                <a href="/posts/{{$post->id}}/edit" class="btn btn-success"> Edit </a>
-            </div>
-            <div class="">
-                {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE', 'class' => 'right-align']) !!}
-                    {!! Form::submit('Delete', ['class' => 'btn btn-primary']) !!}
-                {!! Form::close() !!}        
-            </div>
-        </div>
-
+        @if(!Auth::guest())
+            @if(Auth::user()->id == $post->user_id)
+                <div class="row">
+                    <div class="col s6">
+                        <a href="/posts/{{$post->id}}/edit" class="btn blue"> Edit </a>
+                    </div>
+                    <div class="">
+                        {!! Form::open(['action' => ['PostsController@destroy', $post->id], 'method' => 'DELETE', 'class' => 'right-align']) !!}
+                            {!! Form::submit('Delete', ['class' => 'btn red']) !!}
+                        {!! Form::close() !!}        
+                    </div>
+                </div>
+            @endif
+        @endif
         
     </div>
 @endsection
